+++ skbio/stats/ordination/__init__.py
@@ -51,6 +51,13 @@ Canonical analysis
    rda
 
 
+Graph embedding
+---------------
+
+.. autosummary::
+   :toctree:
+
+   graphembed
 
 Ordination results
 ------------------
@@ -167,6 +174,7 @@ from ._correspondence_analysis import ca
 from ._canonical_correspondence_analysis import cca
 from ._principal_coordinate_analysis import pcoa, pcoa_biplot
 from ._principal_component_analysis import pca
+from ._graphembed import graphembed
 from ._ordination_results import OrdinationResults
 from ._utils import (
     mean_and_std,
@@ -185,6 +193,7 @@ __all__ = [
     # "pca",
     "pcoa",
     "pcoa_biplot",
+    "graphembed",
     "OrdinationResults",
     "mean_and_std",
     "scale",