+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Features
 
+* Added `graphembed` to `skbio.stats.ordination` to support computing representation vectors via the `graphembed_rs` framework, if installed. ([#2212](https://github.com/scikit-bio/scikit-bio/issues/2212))
 * Added support to `TreeNode.from_taxonomy` for parsing taxonomic lineage strings into trees with an optional `extract_rank` parameter. ([#2406](https://github.com/scikit-bio/scikit-bio/pull/2406))
-
 
 ## Version 0.7.2